+++ bms-interface/bms-api/src/main/java/com/bench/bms/api/model/res/WarnVo.java
@@ -9,7 +9,7 @@ import lombok.Data;
 @Data
 public class WarnVo {
 
-    private Integer carId;
+    private Long carId;
 
     private String batteryType;
